requirements/core.txt
fastapi[all]
Flask-SQLAlchemy
psycopg2-binary
pandas
python-multipart
PyDrive2
numpy
